+++ bdhw2/setup.h
@@ -8,15 +8,19 @@
 using std::vector;
 
 struct counterParties {
-	double hazardRate;
-	double netDeal;
+	float hazardRate;
+	float netDeal;
 };
 
+//initialize counterparties with predefined hazard rates
 void setupCounterparties(vector<counterParties>& cp);
+
+//randomly assign deals to counterparties based on specified ratio
 void allocateDeals(vector<counterParties>& cp);
 
 //returns a random deal within the range, can be positive or negative
 long getRandomDeal();
+
 //get a random counter party number (based on predefined allocation ratios)
 long getRandomAllocation();
